Update e correções na parte de cadastro do sistema

+++ src/main/java/br/app/adv/main/person/adv/AdvogadoUserDAO.java
@@ -1,10 +1,13 @@
 package br.app.adv.main.person.adv;
 
+import java.math.BigInteger;
+
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
 public interface AdvogadoUserDAO extends CrudRepository<AdvogadoUser, Long> {
 
-	AdvogadoUser findByPessoaId(long id);
+	AdvogadoUser findByPessoaId(BigInteger id);
+
 }
